backend/requirements.txt
Django==4.2.24
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-ckeditor==6.7.1
Pillow==10.4.0
gunicorn==22.0.0
python-decouple==3.8
dj-database-url==2.2.0
whitenoise==6.6.0
django-jazzmin==3.0.1
